Integrate mongo/ose instead of client-sessions for centralized storage.
Used for sessions, as well as user settings. This will be nice when we have user preferences we need to store, so users don't have to set them independently on all of their devices. Note that we could, theoretically, keep the session data in client-sessions rather than mongo, which is admittedly a poor fit for session data.